+++ RELEASE-NOTES.md
@@ -5,8 +5,11 @@
 **NEW**
 - **Model** - Add `demo` attribute to video data model #1
 - **Feature** - Add demo lessons #2
+- **Model** - Create video collection data models #8
+- **Model** - Add `restricted` attribute to device data model #9
 
 **IMPROVED**
+- **Model** - Change `demo` attribute of video data model to `restricted` #7
 
 **FIXED**